requirements.txt
google-generativeai>=0.3.0
streamlit>=1.28.0
plotly>=5.17.0
networkx>=3.1
python-dotenv>=1.0.0
pytest>=7.4.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
